fixed https://github.com/NuGet/Home/issues/1354

+++ src/NuGet.Core/NuGet.ProjectManagement/Projects/PackagesConfigNuGetProject.cs
@@ -184,7 +184,7 @@ namespace NuGet.ProjectManagement
                 {
                     // Matching packageReference is found and is the only entry
                     // Then just delete the packages.config file 
-                    if (installedPackagesList.Count == 1)
+                    if (installedPackagesList.Count == 1 && nuGetProjectContext.ActionType == NuGetActionType.Uninstall)
                     {
                         FileSystemUtility.DeleteFile(FullPath, nuGetProjectContext);
                     }
